--- src/app/_pages/registroalquiler/page.tsx
@@ -1,9 +1,0 @@
-export default function RegistroAlquiler () {
-
-    return(
-        <div>
-        <h1>Ingrese sus datos por favor</h1>
-        <h1>Por favor</h1>
-        </div>
-    )
-}
