--- src/app/_paginas/page.tsx
@@ -1,20 +1,0 @@
-
-import Inicio from "../[pages]/inicio/inicio";
-import Navbar from "../_components/Navbar";
-
-
-
-export default function Home() {
-  return (
-    <div>
-      <div>
-          <Inicio/>
-          <h1>HOLA</h1>
-          
-      </div>
-    </div>
-  );
-}
-
-
-
